add tile breaking

+++ source/DM_Enemy.cpp
@@ -1,7 +1,7 @@
 // Nate Evans 21144881
 #include "DM_Enemy.h"
 
-DM_Enemy::DM_Enemy(int x, int y, int health, int money, Personality personality)
+DM_Enemy::DM_Enemy(float x, float y, int health, int money, Personality personality)
     : DM_Entity(x, y, health, money)
     , m_personality(personality)
     , m_isAttacking(false)
