+++ source/DM_Enemy.h
@@ -15,7 +15,7 @@ class DM_Enemy : public DM_Entity
 {
     //methods:
 public:
-    DM_Enemy(int x, int y, int health, int money, Personality personality);
+    DM_Enemy(float x, float y, int health, int money, Personality personality);
 
     void tick() override;
