+++ source/DM_Entity.cpp
@@ -1,7 +1,7 @@
 // Nate Evans 21144882
 #include "DM_Entity.h"
 
-DM_Entity::DM_Entity(int x, int y, int health, int money)
+DM_Entity::DM_Entity(float x, float y, int health, int money)
     : xTile(x)
     , yTile(y)
     , health(health)
